Simplified usage of `os::pagesize` in mesos.
Review: https://reviews.apache.org/r/51968/

+++ src/slave/container_loggers/lib_logrotate.hpp
@@ -86,7 +86,7 @@ struct LoggerFlags : public virtual flags::FlagsBase
 
   static Option<Error> validateSize(const Bytes& value)
   {
-    if (value.bytes() < static_cast<uint64_t>(os::pagesize())) {
+    if (value.bytes() < os::pagesize()) {
       return Error(
           "Expected --max_stdout_size and --max_stderr_size of "
           "at least " + stringify(os::pagesize()) + " bytes");
